requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.10.0
click==8.1.7
Django==4.2.9
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
flake8==7.1.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
psycopg2-binary==2.9.6
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
